test: split sync trigger utils to one export per file
