+++ packages/twenty-server/test/integration/calendar/suites/calendar-sync-failure-lifecycle.integration-spec.ts
@@ -23,7 +23,7 @@ import {
   updateCalendarChannel,
 } from 'test/integration/messaging/utils/query-messaging.util';
 import { getCoreRepository } from 'test/integration/utils/get-core-repository.util';
-import { runSyncCron } from 'test/integration/utils/run-sync.util';
+import { runSyncCron } from 'test/integration/utils/run-sync-cron.util';
 
 const THROTTLED_HANDLE = 'calendar-throttled@apple.dev';
 const REVOKED_HANDLE = 'calendar-revoked@apple.dev';
