+++ packages/twenty-server/test/integration/utils/start-channel-sync-and-await.util.ts
@@ -1,0 +1,9 @@
+import { startChannelSync } from 'test/integration/messaging/utils/query-messaging.util';
+import { waitForAllJobsToFinish } from 'test/integration/utils/wait-for-all-jobs-to-finish.util';
+
+export const startChannelSyncAndAwait = async (
+  connectedAccountId: string,
+): Promise<void> => {
+  await startChannelSync(connectedAccountId);
+  await waitForAllJobsToFinish();
+};
